backend/pyproject.toml
[project]
name = "recipe-backend"
version = "0.1.0"
description = "Recipe app backend — FastAPI + SQLite + markitdown"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-multipart>=0.0.20",
    "pillow>=10.0",
]
